Update apps/geoipupdate.c
It's possible that GeoIP_update_database_general() fails on one loop iteration, but then it succeeds on a future iteration.  This would zero out err, so main would end up exiting with a value of 0. 

This change only sets err in the loop if GeoIP_update_database_general() fails, so if any of the updates fail, the program will still return a non-zero value.