libGeoIP: clarify control flow by removing unnecessary 'else'.
If the last statement in an 'if' body is a control flow statement (e.g.
'break', 'continue', 'return', or 'goto') then the 'else' body becomes
unnecessary. That is, all the statements in the 'else' body can be
moved out of the nesting and directly follow the preceeding 'if'
statement. Example:

    if (test) {
        this();
        return;
    } else {
        that();
    }

can be more clearly expressed as:

    if (test) {
        this();
        return;
    }
    that();

instead.