Fix segfault in geoiplookup when passed invalid DB
This fixes the issue reported in:

https://bugzilla.redhat.com/show_bug.cgi?id=1180874

Although this fix acts as a good sanity check, a better fix would
be to not set the database type to GEOIP_COUNTRY_EDITION unless we
first verified that it was of this type. However, such a change may
break compatibility with some databases in the wild, and this is
a legacy product that is in maintenance-only mode.