libgeoip: add GeoIP_free() function
Recently I spent some time on Windows trying to debug a crash
occurring in `free()` caused by the fact that libgeoip was using
the MT release C runtime, while my library was using the debug
version of the same runtime.

While that should be avoidable by making sure you're compiling
with the correct runtime, I think it should also be nice to have
an API that calls the `free()` matching the `malloc()` that has
been used by GeoIP to allocate memory.

This way, one can just use `GeoIP_free()` to free memory that
was allocated by GeoIP without wondering about the runtime that
is used. A similar concept is `freeaddrinfo()` in libc.